pyproject.toml
[tool.poetry]
name = "mlopdracht"
version = "0.1.0"
description = ""
authors = ["R.Grouls <rgrouls@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
torch = { url = "https://download.pytorch.org/whl/cpu/torch-2.0.1%2Bcpu-cp39-cp39-linux_x86_64.whl"}
torchtext = "*"
torch-tb-profiler = "^0.4.0"
torchinfo = "^1.7.1"
torchsummary = "^1.5.1"
pandas = "^1.4.0"
loguru = "^0.7.0"
tqdm = "^4.65.0"
numpy = "^1.24.3"
requests = "^2.30.0"
pydantic = "^1.10.7"
jupyter = "^1.0.0"
ray = {extras = ["tune"], version = "^2.4.0"}
filelock = "^3.12.0"
plotly-express = "^0.4.1"
streamlit = "^1.23.1"
streamlit-option-menu = "^0.3.6"

[tool.poetry.group.tuning.dependencies]
gin-config = "^0.5.0"
protobuf = "^3.20" # needed for ray
bayesian-optimization = "^1.4.2"
hpbandster = "^0.7.4"
configspace = "^0.6.0"
mlflow = "^2.4.0"
tensorboard = "^2.13.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pep8-naming = "^0.13.3"
flake8-annotations = "^3.0.1"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
